update after review (#7622)
* update after review

* cleaning

* typo
