+++ libs/contract/src/lib/term/+state/term.firestore.ts
@@ -1,12 +1,12 @@
-import { MovieLanguageSpecification } from "@blockframes/movie/+state/movie.firestore";
+import { LanguageRecord } from "@blockframes/movie/+state/movie.firestore";
 import type { Media, Territory } from "@blockframes/utils/static-model";
 import type firebase from 'firebase'
 import { Timestamp } from "@blockframes/utils/common-interfaces/timestamp";
 import { staticModel } from "@blockframes/utils/static-model";
 import { format } from "date-fns";
-import { hydrateLanguageForEmail } from '@blockframes/contract/negotiation/utils';
+import { toLanguageVersionString } from "@blockframes/utils/utils";
 
-export function createMailTerm(terms:BucketTerm<Timestamp>[]){
+export function createMailTerm(terms: BucketTerm<Timestamp>[]) {
   return terms.map(term => ({
     ...term,
     territories: term.territories.map(territory => staticModel['territories'][territory]).join(', '),
@@ -15,13 +15,11 @@ export function createMailTerm(terms:BucketTerm<Timestamp>[]){
       from: format(term.duration.from.toDate(), 'dd MMM, yyyy'),
       to: format(term.duration.to.toDate(), 'dd MMM, yyyy'),
     },
-    languages: hydrateLanguageForEmail(term.languages),
+    languages: toLanguageVersionString(term.languages),
   }))
 }
 
-export type MailTerm = ReturnType< typeof createMailTerm>[number]
-
-
+export type MailTerm = ReturnType<typeof createMailTerm>[number];
 
 export interface Duration<T extends Date | firebase.firestore.Timestamp = Date> {
   from: T,
@@ -33,7 +31,7 @@ export interface BucketTerm<T extends Date | firebase.firestore.Timestamp = Date
   duration: Duration<T>;
   territories: Territory[];
   exclusive: boolean;
-  languages: Record<string, MovieLanguageSpecification>;
+  languages: LanguageRecord;
 }
 
 /**
